requirements.txt
motor
discord.py
dnspython